몬스터 동작 추가

+++ Item Data/PortionItemData.cs
@@ -13,8 +13,8 @@ namespace Rito.InventorySystem
     public class PortionItemData : CountableItemData
     {
         /// <summary> 효과량(회복량 등) </summary>
-        public float Value => _value;
-        [SerializeField] private float _value;
+        public int Value => _value;
+        [SerializeField] private int _value;
         public override Item CreateItem()
         {
             return new PortionItem(this);
